==> notebooks/01_eda.ipynb <==
import pandas as pd
import numpy as np
import duckdb
from pathlib import Path
import os

LOADING DATA - *Lending Club dataset 2018-2019*

os.chdir('c:\\Renzo\\Projects\\credit-risk-ai\\')
os.getcwd()

# As this is the exploration phase, we only need to check a sample of the data, so lets pick 1000

con = duckdb.connect()

df = con.execute("""
    SELECT *
    FROM read_csv_auto(
        'data/accepted_2007_to_2018Q4.csv',
        types={'id': 'VARCHAR'}  -- Fix the footer issue
    )
    WHERE issue_d LIKE '%-2017' OR issue_d LIKE '%-2018'
    LIMIT 1000
""").df()

print(df.shape)
df.head()

We are building a supervised learning model. Hence, we need a target. We will focus on loans that we have a real outcome for (i.e. loans that we know whether the borrower payed the loan back or defaulted.) This, we will create a binary target variable to use as a ground truth. In this case, loans with status labeled as **charged off** or **default** will be = 1 (flagged), and the rest will be 0.

df_filtered = con.execute("""
    SELECT 
        CASE WHEN loan_status IN ('Default', 'Charged Off') THEN 1 ELSE 0 END AS target,
        loan_amnt, int_rate, installment, annual_inc, dti, revol_bal, revol_util,
        total_acc, open_acc, pub_rec, fico_range_low, fico_range_high,
        inq_last_6mths, delinq_2yrs, earliest_cr_line, term,
        home_ownership, purpose, grade, sub_grade, emp_length,
        verification_status, addr_state, issue_d
    FROM read_csv_auto(
        'data/accepted_2007_to_2018Q4.csv',
        types={'id': 'VARCHAR'}
    )
    WHERE 
        loan_status IN ('Fully Paid', 'Charged Off', 'Default')
        AND (issue_d LIKE '%-2017' OR issue_d LIKE '%-2018')
        AND annual_inc > 0
        AND dti IS NOT NULL
        AND revol_util IS NOT NULL
""").df()

print(f'Shape of full filtered dataset {df_filtered.shape}')
print(f'% of defaulted or charged off loans: {df_filtered['target'].mean():.2%}')
df.head()


print(df_filtered['issue_d'].str[-4:].value_counts())

Now that we have a ground truth, we will turn the rest of the data (columns) into features that the model can learn from, and drop the rest.

df = df_filtered.copy()

# Creating column: credit_age

df['issue_d'] = pd.to_datetime(df['issue_d'], format='%b-%Y')
df['earliest_cr_line'] = pd.to_datetime(df['earliest_cr_line'], format='%b-%Y')

df['credit_age'] = df['issue_d'] - df['earliest_cr_line']
df['credit_age'] = df['credit_age'].dt.days/365.25
df['credit_age'] = df['credit_age'].clip(lower=0)

# Creating column: payment_to_income

df['payment_to_income'] = (df['installment']*12)/df['annual_inc'].replace(0, np.nan)

# Creating column: revol_utilization_trend

df['revol_utilization_trend'] = df['revol_bal']/(df['annual_inc']+1)

# Creating column: log_annual_inc, log_loan_amnt, log_revol_bal

df['log_annual_inc'] = np.log1p(df['annual_inc'])
df['log_loan_amnt'] = np.log1p(df['loan_amnt'])
df['log_revol_bal'] = np.log1p(df['revol_bal'])

df['term'].value_counts()


# Creating column: term_months

df['term_months'] = df['term'].map({'36 months': 36, '60 months': 60})

# Creating column: fisco_avg

df['fico_avg'] = (df['fico_range_low'] + df['fico_range_high'])/2


cols_to_drop = ['earliest_cr_line', 'term', 'fico_range_low', 'fico_range_high']
df_clean = df.drop(columns=cols_to_drop)

print(f"New shape: {df_clean.shape}")
print(f"New columns: {len(list(df_clean.columns))}")
df_clean[['target', 'credit_age', 'payment_to_income', 'fico_avg', 'term_months']].head()

# Save clean

df_clean.to_parquet('data/processed.parquet', index=False)
