==> notebooks/02_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import duckdb
from pathlib import Path
import os
from sklearn.impute import SimpleImputer

PREPROCESSING

Now that we have selecte the columns we want to use, we need to prepare the data for our the model. This means, we need to transform the columns to numerical data so we are able to train our XGBoost model.

os.chdir('c:\\Renzo\\Projects\\credit-risk-ai\\')

df_clean = pd.read_parquet('data/processed.parquet')

**MISSING VALUE HANLING + ENCODING**

df_clean['emp_length'].value_counts()

df_clean['emp_length'].isna().sum()

emp_map = {
    '< 1 year': 0,
    '1 year': 1,
    '2 years': 2,
    '3 years': 3,
    '4 years': 4,
    '5 years': 5,
    '6 years': 6,
    '7 years': 7,
    '8 years': 8,
    '9 years': 9,
    '10+ years': 10
}

df_clean['emp_length'] = df_clean['emp_length'].map(emp_map).fillna(-1).astype(int)

# Fill all nan values in numeric columns with -1

num_cols = df_clean.select_dtypes(include=np.number).columns

imputer = SimpleImputer(strategy='constant', fill_value=-1)
df_clean[num_cols] = imputer.fit_transform(df_clean[num_cols])

# Transforming grade and subgrade to numerical values

df_clean['grade'].value_counts()

grade_map = {
    'A':1,
    'B':2,
    'C':3,
    'D':4,
    'E':5,
    'F':6,
    'G':7,
}

df_clean['grade_ord'] = df_clean['grade'].map(grade_map)

subgrade_map = {f'{g}{i}': (grade_map[g]-1)*5 + i for g in grade_map for i in range(1,6)}
print(subgrade_map)

df_clean['sub_grade_ord'] = df_clean['sub_grade'].map(subgrade_map)

# One-hot encoding of Home ownership, Purpose and Verification status

df_clean = pd.get_dummies(df_clean, columns=['home_ownership', 'purpose','verification_status'], drop_first=True, dtype=int)

# Creating new temp column for split: issue_year

df_clean['issue_year'] = df_clean['issue_d'].dt.year

train_df = df_clean[df_clean['issue_year'] == 2017].copy()
test_df = df_clean[df_clean['issue_year'] == 2018].copy()

train_df, test_df = train_df.align(test_df, join='left', axis=1, fill_value=0)

state_target = train_df.groupby('addr_state')['target'].mean()
train_df['addr_state_te'] = train_df['addr_state'].map(state_target)
test_df['addr_state_te'] = test_df['addr_state'].map(state_target).fillna(state_target.mean())

train_df.drop(columns=['issue_d', 'issue_year'], inplace=True)
test_df.drop(columns=['issue_d', 'issue_year'], inplace=True)

df_clean = df_clean.drop(columns=['issue_d', 'issue_year'])

final_drop = ['grade', 'sub_grade', 'addr_state']

train_df.drop(columns=final_drop, inplace=True)
test_df.drop(columns=final_drop, inplace=True)
df_clean = df_clean.drop(columns=final_drop)

train_df.shape

train_df.columns

test_df.shape

test_df.columns

df_clean.columns

df_clean.shape

# Save clean
df_clean.to_parquet('data/processed_encoded.parquet', index=False)

# Save final train/test
train_df.to_parquet('data/train_encoded.parquet', index=False)
test_df.to_parquet('data/test_encoded.parquet', index=False)
